requirements.txt
sqlalchemy
alembic
psycopg2-binary
fastapi
pydantic-settings
pydantic
rq
redis
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
chromadb
pymupdf
langchain
langchain-openai
langchain-community
tiktoken
